Fix idempotency issues with git_credentials resource
We've had a long standing issue where the git_credentials is not idempotent.
This is for a few reasons, but primarily due to how the git binary also munges
the file we write to. So instead of using a template, let's use the
append_if_no_line from the line cookbook to ensure the line is there even if it
gets reordered.

Other fixes include:

- Allow adding multiple credentials to the same user with multiple
  git_credentials resources, also add a test to ensure this always works moving
  forward.
- Fix idempotency with :delete action
- Ensure we delete the file with sensitive true
- Add Dokken config
- Fix provisioner block in kitchen.yml
- Migrate to proper library helpers
- Cookstyle fixes
- Adjust tests so that idempotency works correctly
- Cookstyle fixes

Signed-off-by: Lance Albertson <lance@osuosl.org>